refactor: Move GetContribs module to utils folder
This commit moves the GetContribs module from the src/GetRessources folder to the src/utils/GetRessources folder. The module is responsible for fetching the list of contributors from the GitHub API and filtering out team members. This change improves code organization and follows the convention of placing utility modules in the utils folder.